Make scli package pip-compliant
With this scli can be installed via pip or pipx. As of now the
dependency to signal-cli is ignored, meaning that it has to be installed
manually beforehand.

QR code support as well as readline have been marked as extras however
so that the minimal package does not require them.

Fixes: https://github.com/isamert/scli/issues/129